change item path
